+++ packages/app/src/components/adaptive-text-input.tsx
@@ -1,7 +1,5 @@
 import React, { forwardRef, useCallback, useEffect, useRef } from "react";
 import { StyleSheet, withUnistyles } from "react-native-unistyles";
-import { useIsCompactFormFactor } from "@/constants/layout";
-import { isNative } from "@/constants/platform";
 import { createControlGeometry } from "@/components/ui/control-geometry";
 import {
   EditingTextInput,
@@ -29,7 +27,6 @@ const ThemedTextInput = withUnistyles(EditingTextInput, (theme) => ({
 
 export const AdaptiveTextInput = forwardRef<EditingTextInputHandle, AdaptiveTextInputProps>(
   function AdaptiveTextInputInner(props, ref) {
-    const isMobile = useIsCompactFormFactor();
     const { initialValue, resetKey, style, onChangeText, ...inputProps } = props;
     const inputRef = useRef<EditingTextInputHandle | null>(null);
     const replacementTextRef = useRef(initialValue ?? "");
@@ -57,7 +54,6 @@ export const AdaptiveTextInput = forwardRef<EditingTextInputHandle, AdaptiveText
         initialValue={initialValue}
         onChangeText={onChangeText}
         style={[styles.outline, style, styles.text]}
-        variant={isMobile && isNative ? "bottom-sheet" : "default"}
       />
     );
   },
