+++ packages/app/src/components/ui/isolated-bottom-sheet-modal/index.tsx
@@ -9,6 +9,7 @@ import {
   type BottomSheetController,
   createBottomSheetVisibilityTracker,
 } from "./visibility-tracker";
+import { BottomSheetTextInputScope } from "@/components/ui/text-input/bottom-sheet-scope";
 
 type GorhomBottomSheetModalMethods = ElementRef<typeof GorhomBottomSheetModal>;
 
@@ -64,7 +65,9 @@ export const IsolatedBottomSheetModal = forwardRef<
       enableDismissOnClose
       stackBehavior={presentation}
     >
-      {contextBridge ? contextBridge(children) : children}
+      <BottomSheetTextInputScope>
+        {contextBridge ? contextBridge(children) : children}
+      </BottomSheetTextInputScope>
     </GorhomBottomSheetModal>
   );
