+++ packages/app/src/components/ui/text-input/text-input.native.test.tsx
@@ -1,14 +1,43 @@
 // @vitest-environment jsdom
 import React, { act, createRef } from "react";
 import { createRoot, type Root } from "react-dom/client";
-import { afterEach, beforeEach, describe, expect, it } from "vitest";
-import { EditingTextInput } from "./text-input";
+import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
+import { BottomSheetTextInputScope } from "./bottom-sheet-scope";
+import { EditingTextInput } from "./text-input.native";
 import type { EditingTextInputHandle } from "./types";
+
+const bottomSheetTextInputRender = vi.hoisted(() => vi.fn());
+
+vi.mock("@gorhom/bottom-sheet", async () => {
+  const ReactModule = await import("react");
+  return {
+    BottomSheetTextInput: ReactModule.forwardRef<HTMLInputElement, Record<string, unknown>>(
+      (props, ref) => {
+        bottomSheetTextInputRender(props);
+        return ReactModule.createElement("input", {
+          ...props,
+          ref,
+          "data-bottom-sheet-input": true,
+        });
+      },
+    ),
+  };
+});
+
+vi.mock("@mattermost/react-native-paste-input", async () => {
+  const ReactModule = await import("react");
+  return {
+    default: ReactModule.forwardRef<HTMLInputElement, Record<string, unknown>>((props, ref) =>
+      ReactModule.createElement("input", { ...props, ref }),
+    ),
+  };
+});
 
 let container: HTMLDivElement | null = null;
 let root: Root | null = null;
 
 beforeEach(() => {
+  bottomSheetTextInputRender.mockClear();
   container = document.createElement("div");
   document.body.appendChild(container);
   root = createRoot(container);
@@ -28,6 +57,25 @@ afterEach(() => {
 function noop() {}
 
 describe("EditingTextInputNative", () => {
+  it("uses the bottom-sheet input only inside a bottom sheet", () => {
+    act(() => {
+      root?.render(
+        <>
+          <EditingTextInput testID="outside" />
+          <BottomSheetTextInputScope>
+            <EditingTextInput testID="inside" />
+          </BottomSheetTextInputScope>
+        </>,
+      );
+    });
+
+    expect(
+      container?.querySelector('[data-testid="outside"]')?.getAttribute("data-bottom-sheet-input"),
+    ).toBeNull();
+    expect(bottomSheetTextInputRender).toHaveBeenCalledOnce();
+    expect(bottomSheetTextInputRender.mock.calls[0]?.[0]).toMatchObject({ testID: "inside" });
+  });
+
   it("clears text via clear() when replaceText receives an empty string", () => {
     const handleRef = createRef<EditingTextInputHandle>();
